Revert "refacto(client/mapping-sync): move files around and add doc to the crate"
This reverts commit cac504da607df1ad5587888a13a08ccaa2778df0.
