+++ crates/client/db/src/lib.rs
@@ -22,6 +22,7 @@ use std::sync::{Arc, Mutex};
 
 use mapping_db::MappingDb;
 use meta_db::MetaDb;
+// Substrate
 use sc_client_db::DatabaseSource;
 use sp_database::Database;
 use sp_runtime::traits::Block as BlockT;
