+++ crates/node/src/service.rs
@@ -7,7 +7,7 @@ use futures::future;
 use futures::prelude::*;
 use madara_runtime::opaque::Block;
 use madara_runtime::{self, RuntimeApi};
-use mc_mapping_sync::MappingSyncWorker;
+use mc_mapping_sync::{MappingSyncWorker, SyncStrategy};
 use mc_storage::overrides_handle;
 use sc_client_api::{BlockBackend, BlockchainEvents};
 use sc_consensus_aura::{ImportQueueParams, SlotProportion, StartAuraParams};
@@ -259,6 +259,7 @@ pub fn new_full(mut config: Configuration) -> Result<TaskManager, ServiceError> 
             madara_backend.clone(),
             3,
             0,
+            SyncStrategy::Normal,
         )
         .for_each(|()| future::ready(())),
     );
